+++ EduTech/EduTech/src/main/java/com/EduTech/EduTech/repository/ContenidoEduRepository.java
@@ -3,5 +3,5 @@ package com.EduTech.EduTech.repository;
 import com.EduTech.EduTech.model.ContenidoEdu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContenidoEduRepository extends JpaRepository<ContenidoEdu,Long> {
+public interface ContenidoEduRepository extends JpaRepository<ContenidoEdu,Integer> {
 }
